Rename theme card component
